Add BaseBottomSheetDialogFragment support
