+++ app/src/main/java/com/marekpdev/shoppingapp/ui/search/sort/SortBottomSheet.kt
@@ -12,9 +12,11 @@ import androidx.lifecycle.ViewModelProviders
 import com.google.android.material.bottomsheet.BottomSheetBehavior
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 import com.marekpdev.shoppingapp.R
+import com.marekpdev.shoppingapp.databinding.BottomSheetFilterBinding
 import com.marekpdev.shoppingapp.databinding.BottomSheetSortBinding
 import com.marekpdev.shoppingapp.databinding.FragmentSearchBinding
 import com.marekpdev.shoppingapp.mvi.MviView
+import com.marekpdev.shoppingapp.ui.base.BaseBottomSheetDialogFragment
 import com.marekpdev.shoppingapp.ui.search.SearchAction
 import com.marekpdev.shoppingapp.ui.search.SearchCommand
 import com.marekpdev.shoppingapp.ui.search.SearchState
@@ -26,32 +28,11 @@ import java.lang.StringBuilder
  * Created by Marek Pszczolka on 07/06/2022.
  */
 @AndroidEntryPoint
-class SortBottomSheet: BottomSheetDialogFragment(), MviView<SearchState, SearchCommand> {
+class SortBottomSheet: BaseBottomSheetDialogFragment<SearchState, SearchAction, SearchCommand, BottomSheetSortBinding>(R.layout.bottom_sheet_sort) {
 
-    private val viewModel by viewModels<SortBottomSheetViewModel>()
+    override val viewModel by viewModels<SortBottomSheetViewModel>()
 
-    private lateinit var binding: BottomSheetSortBinding
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        binding = DataBindingUtil.inflate(inflater, R.layout.bottom_sheet_sort, container, false)
-        return binding.root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-        binding.apply {
-            lifecycleOwner = this@SortBottomSheet
-            initLayout(this)
-        }
-
-    }
-
-    private fun initLayout(binding: BottomSheetSortBinding) = binding.apply {
-        viewModel.bind(viewLifecycleOwner, this@SortBottomSheet)
+    override fun initLayout(binding: BottomSheetSortBinding) = with(binding) {
 
         BottomSheetBehavior.from(binding.llBottomSheet).state = BottomSheetBehavior.STATE_EXPANDED
